add cache dir
